Resolve cyclic dependendencies (2 files)
- `src/constants/report.constant.ts`
- `src/types/report.type.ts`
