+++ src/components/report-viewer/log-table.jsx
@@ -16,7 +16,7 @@ import { getColumnFilterValue, useAggridRowFilter } from 'hooks/use-aggrid-row-f
 import { LOGS_STORE_FIELD } from 'utils/store-sort-filter-fields';
 import { useReportFetcher } from 'hooks/use-report-fetcher';
 import { useDispatch } from 'react-redux';
-import { getDefaultSeverityFilter } from 'utils/report-severity.utils';
+import { getDefaultSeverityFilter } from '../../utils/report-severity';
 import PropTypes from 'prop-types';
 import { QuickSearch } from './QuickSearch';
 import { Box } from '@mui/material';
