+++ src/hooks/use-report-fetcher.tsx
@@ -11,12 +11,8 @@ import { useCallback, useMemo, useState } from 'react';
 import { fetchNodeReportLogs, fetchParentNodesReport } from '../services/study';
 import { useSnackMessage } from '@gridsuite/commons-ui';
 import { Log, Report, ReportLog, ReportType } from '../types/report.type';
-import { getContainerDefaultSeverityList } from '../utils/report-severity.utils';
-import {
-    COMPUTING_AND_NETWORK_MODIFICATION_TYPE,
-    GLOBAL_REPORT_NODE_LABEL,
-    REPORT_TYPE,
-} from '../constants/report.constant';
+import { getContainerDefaultSeverityList } from '../utils/report-severity';
+import { COMPUTING_AND_NETWORK_MODIFICATION_TYPE, GLOBAL_REPORT_NODE_LABEL } from '../constants/report.constant';
 import { ROOT_NODE_LABEL } from '../constants/node.constant';
 import { mapReportLogs } from '../utils/report-log.mapper';
 
@@ -118,7 +114,7 @@ export const useReportFetcher = (
     const fetchReportLogs = useCallback(
         (reportId: string, severityList: string[], reportType: ReportType, messageFilter: string) => {
             let fetchPromise: (severityList: string[], reportId: string) => Promise<ReportLog[]>;
-            if (reportType === REPORT_TYPE.GLOBAL) {
+            if (reportType === ReportType.GLOBAL) {
                 fetchPromise = (severityList: string[]) =>
                     fetchNodeReportLogs(studyUuid, currentNode!.id, null, severityList, messageFilter, true);
             } else {
